Separa o calendario em página e componente
