+++ frontend/src/App.js
@@ -4,7 +4,7 @@ import { Switch, Route, BrowserRouter } from 'react-router-dom'
 import Login from './pages/login.js'
 import Cadastro from './pages/cadastro.js'
 import MainNavbar from './componentes/MainNavbar.js'
-import Calendario from './pages/calendario.js'
+import PaginaCalendario from './pages/paginacalendario.js'
 import './App.css';
 
 class App extends Component {
@@ -16,7 +16,7 @@ class App extends Component {
                     <Route path="/" exact={true} component={Login} />
                     <Route path="/cadastro" exact={true} component={Cadastro} />
                     <Route path="/MainNavbar" exact={true} component={MainNavbar} />
-                    <Route path="/calendario" exact={true} component={Calendario} />
+                    <Route path="/calendario" exact={true} component={PaginaCalendario} />
                 </Switch> 
             </BrowserRouter>
         );
